add post rotate callbacks
Added a new exported function `postrotate` that can be used to register a callback function to be invoked with the rotated file name just after the current log file is rotated. The file name of the rotated file is passed as an argument. The function is blocking and so any lengthy operation that needs to be done should be done asynchronously. Callbacks can be registered with both RollingFileWriter and RollingLogger instances. E.g.:

```
julia> postrotate(logger) do rotatedfile
           # e.g. code to upload file to permanent store
           # ...
       end
```